decide whether to build fat runtime on non-Linux
The current checking leaves fat runtime support enabled on non-Linux
platforms. For example, on Debian GNU/kFreeBSD (Debian with a
FreeBSD kernel) the current 4.4 release fails to build. For more
details see the failed build log at: http://deb.li/mJce
This patch ensures that the decision whether to build a fat runtime or
not is also made on non-Linux systems, fixing this build.

Signed-off-by: Sascha Steinbiss <sascha@steinbiss.name>